+++ src/application/domain/comment/use_cases/delete_comment.py
@@ -1,5 +1,6 @@
 import logging
 
+from application.core.exceptions.auth_exceptions import AccessDeniedException
 from application.core.exceptions.database_exceptions import (
     CommentNotFoundException,
 )
@@ -21,6 +22,12 @@ class DeleteCommentUseCase:
         self._repo = CommentRepository()
 
     async def execute(self, id: int, current_user: UserResponseSchema) -> None:
+        if not current_user.is_admin:
+            error = AccessDeniedException()
+            logger.error(
+                f'Доступ запрещен: пользователь {current_user.username} попытался удалить комментарий с id {id}'
+            )
+            raise error
         async with self._database.session() as session:
             try:
                 await self._repo.delete(session=session, id=id)
