+++ src/application/domain/location/use_cases/delete_location.py
@@ -1,6 +1,7 @@
 import logging
 import uuid
 
+from application.core.exceptions.auth_exceptions import AccessDeniedException
 from application.core.exceptions.database_exceptions import (
     LocationNotFoundException,
 )
@@ -24,6 +25,12 @@ class DeleteLocationUseCase:
     async def execute(
         self, id: uuid.UUID, current_user: UserResponseSchema
     ) -> None:
+        if not current_user.is_admin:
+            error = AccessDeniedException()
+            logger.error(
+                f'Доступ запрещен: пользователь {current_user.username} попытался удалить местоположение с id {id}'
+            )
+            raise error
         async with self._database.session() as session:
             try:
                 await self._repo.delete(session=session, id=id)
